Extract normalizeRole helper
