+++ src/app/api/admin/users/route.ts
@@ -6,24 +6,7 @@ import { ensureSessionRole, requireApiSession } from "@/app/api/_utils/require-a
 import prisma from "@/lib/prisma";
 import { auth } from "@/lib/auth";
 
-/**
- * Normaliza strings de rol que puedan venir del front.
- * Acepta "comercial" como sinónimo de "usuario" (compatibilidad vieja).
- */
-function normalizeRole(
-  input: string | null | undefined,
-): DbRole | undefined {
-  if (!input) return undefined;
-  const v = input.toLowerCase().trim();
-  if (v === "comercial") return DbRole.usuario;
-  if (v === "usuario") return DbRole.usuario;
-  if (v === "lider") return DbRole.lider;
-  if (v === "admin") return DbRole.admin;
-  if (v === "superadmin") return DbRole.superadmin;
-  return undefined;
-}
-
-export { normalizeRole };
+import { normalizeRole } from "./normalize-role";
 
 export async function GET() {
   const { session, response } = await requireApiSession();
